Add files via upload
Alumni Tracking System is an online-based application that helps to enhance the tracking of 
college graduates. The project primarily aims to replace current tracking procedure of college 
graduates and providing alumni data to college faculties. It aims at developing a web portal 
which will be useful for the college to monitor the alumni's and for the alumni to update their 
status and get notified about the college activities. This project which stores the Alumni 
database which makes easier to maintain records of the alumni along with staff details. The 
proposed system is online so it can be accessed by Alumni anywhere. Handling all records all 
the Records is a difficult task for a user this product provides easy way for maintaining 
records.